Refactor past detention requests into a component
